allow strict dynamic to be injected into CSP header regardless of internally stored policy
